Fixed deleting server and deploying openmrs-webapp snapshot

+++ maven-plugin/src/main/java/org/openmrs/maven/plugins/Delete.java
@@ -1,14 +1,11 @@
 package org.openmrs.maven.plugins;
 
 import org.apache.commons.io.FileUtils;
-import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
 import org.openmrs.maven.plugins.model.Server;
 import org.openmrs.maven.plugins.utility.DBConnector;
-import org.openmrs.maven.plugins.utility.Wizard;
 
-import java.io.File;
 import java.io.IOException;
 import java.sql.SQLException;
 
@@ -28,16 +25,19 @@ public class Delete extends AbstractTask{
     public void executeTask() throws MojoExecutionException, MojoFailureException {
         serverId = wizard.promptForExistingServerIdIfMissing(serverId);
         try {
-            Server props = Server.loadServer(serverId);
-            FileUtils.deleteDirectory(props.getServerDirectory());
-            String dbName = props.getParam(Server.PROPERTY_DB_NAME);
-            String dbUser = props.getParam(Server.PROPERTY_DB_USER);
-            String dbPass = props.getParam(Server.PROPERTY_DB_PASS);
-            String dbUri = props.getParam(Server.PROPERTY_DB_URI);
-            DBConnector connector = new DBConnector(dbUri, dbUser, dbPass, dbName);
-            connector.dropDatabase();
-            connector.close();
-            getLog().info(String.format(TEMPLATE_SUCCESS, props.getServerId()));
+            Server server = Server.loadServer(serverId);
+            FileUtils.deleteDirectory(server.getServerDirectory());
+
+            if (server.isMySqlDb()) {
+                String dbName = server.getParam(Server.PROPERTY_DB_NAME);
+                String dbUser = server.getParam(Server.PROPERTY_DB_USER);
+                String dbPass = server.getParam(Server.PROPERTY_DB_PASS);
+                String dbUri = server.getParam(Server.PROPERTY_DB_URI);
+                DBConnector connector = new DBConnector(dbUri, dbUser, dbPass, dbName);
+                connector.dropDatabase();
+                connector.close();
+            }
+            getLog().info(String.format(TEMPLATE_SUCCESS, server.getServerId()));
         } catch (IOException e) {
             throw new MojoExecutionException(e.getMessage());
         } catch (SQLException e) {
